bruinhub-backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
psycopg==3.1.12
python-dotenv==1.0.0
APScheduler==3.10.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.3.1